More key focus tweaks
This is getting really questionable now. The focus check in
MainWindow can be moved after the OSX special case because on
OSX the text fields do consume the events before MainWindow
gets them. This is not the case on Win32 and I don't know which
one is really the bug, or if some focus setting needs to be
changed in my code.