Process XInput2 events using a native event filter
This is a workaround for the unstable coordinates in the Qt tablet
events (QTBUG-45375). I'm still unsure whether this is a bug in X.Org
or if the 16.16 fixed point values are inherently too imprecise; but
the event filter is able to cleanly replace them with the high
precision axis values without otherwise disturbing event processing.