microdab: Add a subpixel mode to MyPaintTool for strokes less than 2 pixels wide
For dabs with a radius less than 1 the dab doesn't reliably touch
the center of each pixel, to accommodate this the micro_dab() kernels
find some point inside the pixel that does intersect the dab and use
that to approximate the dab's coverage of the pixel.

This logic is heavily inspired by MyPaint's anti-alised dab algorithm,
the major difference being that this calculates everything in the dab's
coordinate space (where 0,0 is the center of the dab). This makes the
near point more accurate but the far point less so.