native-events: Update native event filter to deal with QTBUG-56263
Because newer version of Qt use XInput for all events it's no longer possible
to just intercept the tablet part of the mouse event stream. Rather than try
to fully replicate event delivery chain this version only intercepts events
that are destined for the "mainCanvas" widget.

It also adds a check to avoid an older issue of bypassing modal widgets.